requirements.txt
google-adk
google-genai
pandas
scikit-learn
numpy
gunicorn
uvicorn
fastapi